refactor: Order 도메인 > reader, writer 도메인 레이어로 이동
