+++ commerce-api/src/main/kotlin/com/hanghae/commerce/claim/domain/OrderCancelService.kt
@@ -2,7 +2,7 @@ package com.hanghae.commerce.claim.domain
 
 import com.hanghae.commerce.claim.domain.command.OrderCancelCommand
 import com.hanghae.commerce.event.CommerceEventPublisher
-import com.hanghae.commerce.order.application.OrderWriter
+import com.hanghae.commerce.order.domain.OrderWriter
 import com.hanghae.commerce.order.domain.Order
 import org.springframework.stereotype.Service
